- Fixed some scaling errors in CollideXY();
- Made CollideZ() somewhat better.  Still work to do though.


- Incorporated a shift in the slip angle that depends on
road camber angle for tyres, according to the advice of
Pacejka in 'Tyre and Vehicle Dynamics'. The difference
in handling is small. The important thing is that
now you can 'preload' the tyres so that the car
responds more sharply or more gently.  This is an effect
similar to that obtained by toeing the tyres, but it
is speed dependent.


- Fixed the OPTIMAL model for aerodynamics. Now it is the
default model in PRO mode. It is slightly more efficient
than the other model, so that should make up a bit for the
reduced downforce. I think I will also add an 'efficiency'
term here to model the viscous drag.